test(pg): the search_brain relational retrieval live suite requires the test database instead of skipping, fixture split out (#878)
The live half of the relational retrieval eval read
OPENBRAIN_TEST_DATABASE_URL itself and swapped in describe.skip when it was
unset. That reports 0 pass, N skip, 0 fail and exits 0, which at the exit code
is indistinguishable from a suite that ran and passed, so the Postgres graph
predicates nobody exercised could drift under a green CI.

The 1095-line file is now four:

- search-brain-relational-retrieval-fixture-data.ts holds the fixture types,
  entries, entities, relational questions, and links. No test, no pool.
- search-brain-relational-retrieval-fixture.ts holds the query parser, the
  keyword baseline, the graph oracle, and the mock search pool. No test, no
  real pool.
- search-brain-relational-retrieval.test.ts keeps the 12 unit its, unchanged in
  name, order, and assertion, importing from the two fixture modules.
- search-brain-relational-retrieval.pg.test.ts is a plain describe ending
  "(live Postgres)". It builds its pool at module scope from
  requireTestDatabaseUrl(), so an absent variable throws test_database_required
  and takes the run down loudly.

The pre-commit gate lints staged files whole, so the pre-existing findings in
the touched files are fixed here without disable comments: every it body is
hoisted to a module-scope function called as it("<same name>", fn), the three
explicit any become unknown narrowed at the use site, parseFixtureQuery sheds
its outgoing half into parseOutgoingFixtureQuery, and relationalDbCandidates
sheds its five join builders into relationalJoins. Assertion text and test
names move verbatim.

scripts/assert-db-tests-ran.ts already carries the live describe name with
minTests 2; JUnit confirms the suite emits exactly 2, so the manifest and the
floor stand unchanged.

Done-means: scripts/done-means/878-pg-tests-require-database.sh
